refactoring for tests ver 2.0
