+++ src/test/java/com/epam/seleniumhw/mailru/tests/EmailCompareDraftFieldsTest.java
@@ -1,17 +1,12 @@
 package com.epam.seleniumhw.mailru.tests;
 
 
-import com.epam.seleniumhw.mailru.pageobject.MainPage;
 import com.epam.seleniumhw.mailru.utils.TestDataProvider;
 import io.qameta.allure.Epic;
 import io.qameta.allure.Feature;
 import io.qameta.allure.Story;
 import io.qameta.allure.testng.Tag;
 import org.testng.annotations.Test;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 @Epic("Selenium_HW")
 @Feature("Test_MailRu_Box")
