+++ src/test/java/com/epam/seleniumhw/mailru/tests/EmailLoginTest.java
@@ -1,14 +1,11 @@
 package com.epam.seleniumhw.mailru.tests;
 
-import com.epam.seleniumhw.mailru.pageobject.MainPage;
 import io.qameta.allure.Epic;
 import io.qameta.allure.Feature;
 import io.qameta.allure.Story;
 import io.qameta.allure.testng.Tag;
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 
 @Epic("Selenium_HW")
