+++ src/test/java/com/epam/seleniumhw/mailru/utils/NumberGenerator.java
@@ -4,7 +4,6 @@ import java.util.Random;
 
 public class NumberGenerator {
 
-
     public static String getRandomFirstDomainLvl() {
         StringBuilder init = new StringBuilder();
         Random r = new Random();
